feat(extension): 🎸 video
