+++ packages/extensions/video/README.md
@@ -1,0 +1,13 @@
+# @yiitap/extension-video
+[![Version](https://img.shields.io/npm/v/@yiitap/extension-video.svg?label=version)](https://www.npmjs.com/package/@yiitap/extension-video)
+[![Downloads](https://img.shields.io/npm/dm/@yiitap/extension-video.svg)](https://npmcharts.com/compare/yiitap?minimal=true)
+[![License](https://img.shields.io/npm/l/@yiitap/extension-video.svg)](https://github.com/yiitap/yiitap/blob/main/LICENSE)
+
+## Introduction
+Yiitap is a WYSIWYG rich-text block-based editor built on top of [tiptap](https://tiptap.dev).
+
+## Official Documentation
+Documentation can be found on the [Yiitap website](https://yiitap.pileax.ai).
+
+## License
+Yiitap is open sourced software licensed under the [MIT license](https://github.com/yiitap/yiitap/blob/main/LICENSE).
