+++ packages/extensions/video/src/__tests__/index.spec.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from 'vitest'
+import { mount } from '@vue/test-utils'
+
+describe('Sample Test Suite', () => {
+	it('mounts', () => {
+		console.log('just a sample test')
+	})
+})
