+++ packages/extensions/video/src/index.ts
@@ -1,0 +1,5 @@
+import { Video } from './video'
+
+export * from './video'
+
+export default Video
